+++ src/enbrief.c
@@ -189,7 +189,7 @@ void blinkWidget(MenuItem *item) {
     int16 fromColor;
     if (item->state == 0) {
         item->state = 1;
-        fromColor = (unsigned)item->colorPair >> 4;
+        fromColor = (uint16)item->colorPair >> 4;
         toColor = item->colorPair & 0xF;
         if (item->colorPair != 0) {
             menuLabelSwitch(item, fromColor, toColor);
@@ -197,7 +197,7 @@ void blinkWidget(MenuItem *item) {
     } else {
         item->state = 0;
         fromColor = item->colorPair & 0xF;
-        toColor = (unsigned)item->colorPair >> 4;
+        toColor = (uint16)item->colorPair >> 4;
     }
     if (item->colorPair != 0) {
         menuLabelSwitch(item, fromColor, toColor);
@@ -269,7 +269,7 @@ int16 isPointInRect(const MenuItem *p) {
                 fromColor = colorTablePtr[colorAnimIdx + 1] & 0xF;
                 menuLabelSwitch(menuItem, toColor, fromColor);
                 colorAnimIdx++;
-                colorAnimIdx = (unsigned)colorAnimIdx % *colorTablePtr;
+                colorAnimIdx = (uint16)colorAnimIdx % *colorTablePtr;
             }
         }
 
@@ -628,7 +628,7 @@ static void drawFullPathLines(uint16 maxRecord) {
     int16 curX, curY, prevX, prevY, recIdx;
     prevX = prevY = 0;
     recIdx = -1;
-    while (++recIdx, (flightRecords[recIdx].status & STATUS_TYPE_MASK) != 0 && (unsigned)recIdx <= maxRecord) {
+    while (++recIdx, (flightRecords[recIdx].status & STATUS_TYPE_MASK) != 0 && (uint16)recIdx <= maxRecord) {
         gfx_setColor(0);
         if (recIdx == 0) {
             plotMapPoint(flightRecords[0].mapX, flightRecords[0].mapY, 0, 0);
@@ -645,9 +645,9 @@ static void drawFullPathLines(uint16 maxRecord) {
 }
 
 /* The event map markers for records up to maxRecord. */
-static void drawPathSprites(unsigned int maxRecord) {
-    int recIdx = -1;
-    while (++recIdx, (flightRecords[recIdx].status & STATUS_TYPE_MASK) != 0 && (unsigned)recIdx <= maxRecord) {
+static void drawPathSprites(uint16 maxRecord) {
+    int16 recIdx = -1;
+    while (++recIdx, (flightRecords[recIdx].status & STATUS_TYPE_MASK) != 0 && (uint16)recIdx <= maxRecord) {
         if ((flightRecords[recIdx].status & STATUS_TYPE_MASK) != EVENT_TIMESTAMP)
             drawEventSprite(recIdx);
     }
@@ -760,13 +760,13 @@ char *formatFlightTime(int16 timeValue, char *buffer) {
     }
     timeValue += (miscBits & 0xF) << 8;
     mystrcpy(buffer, "00:00:00");
-    hours = (unsigned)timeValue / 1800;
+    hours = (uint16)timeValue / 1800;
     buffer[0] += nightMission + 1;
     buffer[1] += hours % 10;
-    minutes = ((unsigned)timeValue / 30) % 60;
+    minutes = ((uint16)timeValue / 30) % 60;
     buffer[3] += minutes / 10;
     buffer[4] += minutes % 10;
-    seconds = ((unsigned)timeValue * 2) % 60;
+    seconds = ((uint16)timeValue * 2) % 60;
     buffer[6] += seconds / 10;
     buffer[7] += seconds % 10;
     return buffer;
@@ -786,10 +786,10 @@ void plotMapPoint(int16 x, int16 y, int16 color, int16 unused) {
     sx = mapToScreenX(x);
     sy = mapToScreenY(y);
     if (color != -1 &&
-        (unsigned)sx >= (unsigned)mapViewX1 &&
-        (unsigned)sx < (unsigned)mapViewX2 &&
-        (unsigned)sy >= (unsigned)mapViewY1 &&
-        (unsigned)sy < (unsigned)mapViewY2) {
+        (uint16)sx >= (uint16)mapViewX1 &&
+        (uint16)sx < (uint16)mapViewX2 &&
+        (uint16)sy >= (uint16)mapViewY1 &&
+        (uint16)sy < (uint16)mapViewY2) {
         drawMapPixel(sx, sy, color);
     }
 }
@@ -850,7 +850,7 @@ long calcMissionScore(int16 param) {
         weaponCount = 15;
     }
 
-    for (recIdx = 0; (unsigned)recIdx <= (unsigned)param && flightRecords[recIdx].status; recIdx++) {
+    for (recIdx = 0; (uint16)recIdx <= (uint16)param && flightRecords[recIdx].status; recIdx++) {
         unitId = flightRecords[recIdx].unitId;
         switch (flightRecords[recIdx].status & STATUS_TYPE_MASK) {
         case EVENT_EJECTED:
@@ -978,16 +978,16 @@ void showEventPopup(void) {
         spriteIdx = 10;
         break;
     }
-    if ((unsigned)(mapToScreenX(flightRecords[curRecordIdx].mapX) + mapViewX1) < 115 &&
-        (unsigned)(mapToScreenY(flightRecords[curRecordIdx].mapY) + mapViewY1) < 89) {
+    if ((uint16)(mapToScreenX(flightRecords[curRecordIdx].mapX) + mapViewX1) < 115 &&
+        (uint16)(mapToScreenY(flightRecords[curRecordIdx].mapY) + mapViewY1) < 89) {
         popupX = mapToScreenX(flightRecords[curRecordIdx].mapX) + mapViewX1 + 10;
         popupY = mapToScreenY(flightRecords[curRecordIdx].mapY) + mapViewY1 + 10;
-    } else if ((unsigned)(mapToScreenX(flightRecords[curRecordIdx].mapX) + mapViewX1) >= 115 &&
-               (unsigned)(mapToScreenY(flightRecords[curRecordIdx].mapY) + mapViewY1) < 89) {
+    } else if ((uint16)(mapToScreenX(flightRecords[curRecordIdx].mapX) + mapViewX1) >= 115 &&
+               (uint16)(mapToScreenY(flightRecords[curRecordIdx].mapY) + mapViewY1) < 89) {
         popupX = mapToScreenX(flightRecords[curRecordIdx].mapX) + mapViewX1 - 58;
         popupY = mapToScreenY(flightRecords[curRecordIdx].mapY) + mapViewY1 + 10;
-    } else if ((unsigned)(mapToScreenX(flightRecords[curRecordIdx].mapX) + mapViewX1) >= 115 &&
-               (unsigned)(mapToScreenY(flightRecords[curRecordIdx].mapY) + mapViewY1) >= 89) {
+    } else if ((uint16)(mapToScreenX(flightRecords[curRecordIdx].mapX) + mapViewX1) >= 115 &&
+               (uint16)(mapToScreenY(flightRecords[curRecordIdx].mapY) + mapViewY1) >= 89) {
         popupX = mapToScreenX(flightRecords[curRecordIdx].mapX) + mapViewX1 - 58;
         popupY = mapToScreenY(flightRecords[curRecordIdx].mapY) + mapViewY1 - 40;
     } else {
